feat: change price from Long to Double in Item and ItemDbModel

+++ core/domain/src/main/kotlin/com/metes/worthit/core/domain/entity/Item.kt
@@ -6,7 +6,7 @@ import java.time.LocalDate
 data class Item(
     val id: Int = DEFAULT_ID,
     val name: String,
-    val price: Long?,
+    val price: Double?,
     val currency: Currency,
     val createdAt: Instant,
     val dateOfPurchase: LocalDate,
